근무현황 요약에 승인된 결재문서에 따라휴가, 연장, 야간 근무 반영 안되던거 수정, 잔여휴가 소수점 표현되게 수정, 전자결재 버그 수정

+++ ggirick_client_backend/src/main/java/com/kedu/ggirick_client_backend/controllers/hr/VacationController.java
@@ -26,8 +26,8 @@ public class VacationController {
 
     // 잔여 휴가 조회
     @GetMapping("/remaining")
-    public ResponseEntity<Integer> getRemainingVacation(@AuthenticationPrincipal UserTokenDTO userInfo) {
-        int remaining = vacationService.getRemainingVacation(userInfo.getId());
+    public ResponseEntity<Double> getRemainingVacation(@AuthenticationPrincipal UserTokenDTO userInfo) {
+        double remaining = vacationService.getRemainingVacation(userInfo.getId());
         return ResponseEntity.ok(remaining);
     }
