+++ ggirick_client_backend/src/main/java/com/kedu/ggirick_client_backend/services/hr/VacationService.java
@@ -315,7 +315,7 @@ public class VacationService {
     }
 
     // 잔여휴가 조회
-    public int getRemainingVacation(String employeeId) {
+    public double getRemainingVacation(String employeeId) {
         return annualLeaveGrantService.getRemainingVacation(employeeId);
     }
 }
